backend/pyproject.toml
[project]
name = "rappi-analytics-backend"
version = "0.1.0"
description = "NL-to-SQL analytics chatbot for Rappi operations"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "anthropic",
    "duckdb",
    "sqlglot",
    "pydantic-settings",
    "sse-starlette",
    "jinja2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
